Getting polished: toUppaal, toDot, Show, *, ||, sync.

+++ src/main/scala/ifta/ClockCons.scala
@@ -17,10 +17,10 @@ sealed trait ClockCons {
 }
 
 case object CTrue                extends ClockCons
-case class LT(c:String,n:Float)  extends ClockCons
-case class GT(c:String,n:Float)  extends ClockCons
-case class LE(c:String,n:Float)  extends ClockCons
-case class GE(c:String,n:Float)  extends ClockCons
+case class LT(c:String,n:Int)  extends ClockCons
+case class GT(c:String,n:Int)  extends ClockCons
+case class LE(c:String,n:Int)  extends ClockCons
+case class GE(c:String,n:Int)  extends ClockCons
 case class CAnd(cc1:ClockCons,cc2:ClockCons) extends ClockCons
